+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Do not push the `internalSiteSearchView` event while the `facets` query is not completed.
+
 ## [2.100.0] - 2020-08-11
 ### Added
 - `orderBy`, `page` and filter selection changes to trigger `pageInfo` event.